Switch to TV login method, Poetry
TIDAL changed things around and the old login method doesn't seem to be
working any more. Ported the Android TV OAUTH2 login flow over from
Dniel97's RedSea fork (https://github.com/Dniel97/RedSea), and we're
back in business. It's not as nice as supplying a username and password
directly, but at least we won't have to deal with captchas. A big thanks
to Daniel and the rest of the community for the code and research.

Also switch from setup.py to Poetry for dependency management.